Separate browse object from conductor
