+++ platform/commonUI/browse/res/templates/browse-object.html
@@ -66,5 +66,4 @@
             </mct-representation>
         </div>
     </div>
-    <mct-include key="'conductor'" class="abs holder flex-elem flex-fixed l-flex-row l-time-conductor-holder"></mct-include>
 </div>
