--- src/adapter/vue-adapter/install.js
@@ -1,16 +1,0 @@
-define([
-    './main-adapter',
-    './tree-adapter',
-    './inspector-adapter'
-], function (
-    MainAdapter,
-    TreeAdapter,
-    InspectorAdapter
-) {
-
-    return function install(layout, openmct) {
-        let main = new MainAdapter(layout, openmct);
-        let tree = new TreeAdapter(layout, openmct);
-        let inspector = new InspectorAdapter(layout, openmct);
-    }
-});
