--- src/adapter/vue-adapter/tree-adapter.js
@@ -1,15 +1,0 @@
-define([
-
-], function (
-
-) {
-
-    class TreeAdapter {
-        constructor(layout, openmct) {
-
-        }
-    }
-
-    return TreeAdapter;
-
-});
