+++ src/ui/components/layout/Conductor.vue
@@ -1,0 +1,15 @@
+<template>
+    <div></div>
+</template>
+
+<style lang="scss">
+</style>
+
+<script>
+export default {
+    <mct-include key="'conductor'" class="abs holder flex-elem flex-fixed l-flex-row l-time-conductor-holder"></mct-include>
+
+}
+</script>
+
+
