--- src/ui/components/layout/MctMain.vue
@@ -1,21 +1,0 @@
-<template>
-    <mct-representation key='browse-object'></mct-representation>
-</template>
-
-<style lang="scss">
-.MCT_Main {
-    position: absolute;
-    right: 100px;
-    top: 0px;
-    bottom: 20px;
-    left: 100px;
-    background: green;
-}
-</style>
-
-<script>
-export default {
-}
-</script>
-
-
