+++ src/ui/components/legacy/mct-template.vue
@@ -1,0 +1,38 @@
+<template>
+    <div></div>
+</template>
+
+<script>
+export default {
+    inject: ['openmct'],
+    props: {
+        templateKey: String
+    },
+    mounted() {
+        let openmct = this.openmct;
+        let $injector = openmct.$injector;
+        let angular = openmct.$angular;
+
+        let templateLinker = $injector.get('templateLinker');
+
+        let templateMap = {};
+        $injector.get('templates[]').forEach((t) => {
+            templateMap[t.key] = templateMap[t.key] || t;
+        });
+
+        let $rootScope = $injector.get('$rootScope');
+        this.$scope = $rootScope.$new();
+        console.log('mounting', this.templateKey);
+        console.log('template:', templateMap[this.templateKey]);
+
+        templateLinker.link(
+            this.$scope,
+            angular.element(this.$el),
+            templateMap[this.templateKey]
+        );
+    },
+    destroyed() {
+        this.$scope.$destroy();
+    }
+}
+</script>
